+++ stac_fastapi/types/stac_fastapi/types/core.py
@@ -398,6 +398,25 @@ class BaseCoreClient(LandingPageMixin, abc.ABC):
                 }
             )
 
+        # Add Aggregation links
+        if self.extension_is_enabled("AggregationExtension"):
+            landing_page["links"].extend(
+                [
+                    {
+                        "rel": "aggregate",
+                        "type": "application/json",
+                        "title": "Aggregate",
+                        "href": urljoin(base_url, "aggregate"),
+                    },
+                    {
+                        "rel": "aggregations",
+                        "type": "application/json",
+                        "title": "Aggregations",
+                        "href": urljoin(base_url, "aggregations"),
+                    },
+                ]
+            )
+
         # Add Collections links
         collections = self.all_collections(request=kwargs["request"])
 
@@ -600,6 +619,25 @@ class AsyncBaseCoreClient(LandingPageMixin, abc.ABC):
                     "href": urljoin(base_url, "queryables"),
                     "method": "GET",
                 }
+            )
+
+        # Add Aggregation links
+        if self.extension_is_enabled("AggregationExtension"):
+            landing_page["links"].extend(
+                [
+                    {
+                        "rel": "aggregate",
+                        "type": "application/json",
+                        "title": "Aggregate",
+                        "href": urljoin(base_url, "aggregate"),
+                    },
+                    {
+                        "rel": "aggregations",
+                        "type": "application/json",
+                        "title": "Aggregations",
+                        "href": urljoin(base_url, "aggregations"),
+                    },
+                ]
             )
 
         # Add Collections links